refactor: rename esquery engine to ast engine for better clarity
