+++ src/analyze/helpers/schemas.ts
@@ -50,6 +50,6 @@ export const ruleSchema = z.object({
     description: z.string(),
     tech: z.array(z.literal("next")),
     severity: z.enum(["info", "low", "medium", "high"]),
-    type: z.enum(["request", "esquery"]),
+    type: z.enum(["request", "ast"]),
     steps: z.array(stepSchema),
 });
